Added capability to serve custom notes over a web browser to all students. Added wrapper scripts to protect existing clusters and make it easier to run different commands

+++ delete_cluster.sh
@@ -1,4 +1,13 @@
 ##As a convention USERID will be the current dir name
 export USERID=${PWD##*/}
-
-terraform -chdir=provision/ destroy --auto-approve -var-file ../$USERID.tfvars
+export CLUSTER_NAME=$USERID-dka100
+echo Are you sure you want to delete the cluster? Enter cluster name $CLUSTER_NAME to confirm deletion:
+read CONFIRM
+echo $CONFIRM
+if [[ $CONFIRM == $CLUSTER_NAME ]]; 
+then
+   echo Deleting Cluster $CLUSTER_NAME
+   terraform -chdir=provision/ destroy --auto-approve -var-file ../$USERID.tfvars
+else
+   echo "Skipping"
+fi
